Update uno.py
Howdy @JMTNTBANG

I have had a look at this bug and it seemed to me that the loop was running again causing the message to print twice. I have made a pull request for my update to the Say Uno code which is just adding a boolean and breaking the loops.

Thanks, Tikkitai